style: separate test packages
