+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,9 +1,8 @@
-package racingcar.implementtest;
+package racingcar.domain;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Car;
 
 public class CarTest {
     private Car testCar;
