+++ src/test/java/racingcar/validator/RoundNumberValidatorTest.java
@@ -1,9 +1,8 @@
-package racingcar.implementtest;
+package racingcar.validator;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racingcar.validator.RoundNumberValidator;
 
 public class RoundNumberValidatorTest {
