+++ src/test/java/racingcar/view/InputTest.java
@@ -1,4 +1,4 @@
-package racingcar.implementtest;
+package racingcar.view;
 
 import camp.nextstep.edu.missionutils.Console;
 import java.io.ByteArrayInputStream;
@@ -7,7 +7,6 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import racingcar.view.Input;
 
 public class InputTest {
     private Input input;
